regular commit348

+++ main/accounts/views.py
@@ -53,20 +53,23 @@ def logout_view(request):
 @api_view(['GET','PUT', 'DELETE'])
 def user_actions(request, username):
     user_instance = get_object_or_404(User, username=username)
-
-    if request.method == "GET":
-        serializer = UserSerializer(user_instance)
-        return Response(serializer.data, status=status.HTTP_200_OK)
-    elif request.method == "PUT":
-        serializer = UserSerializer(user_instance, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
+    if request.user == user_instance:
+        if request.method == "GET":
+            serializer = UserSerializer(user_instance)
             return Response(serializer.data, status=status.HTTP_200_OK)
-        else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    elif request.method == "DELETE":
-        user_instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        elif request.method == "PUT":
+            serializer = UserSerializer(user_instance, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        elif request.method == "DELETE":
+            user_instance.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        
+    else:
+        return Response(status=status.HTTP_403_FORBIDDEN)
